pyproject.toml
[tool.poetry]
name = "kinogram"
version = "0.1.0"
description = ""
authors = ["kilzzze <bezkrovnyy2012@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "^3.1.1"
python-dotenv = "^1.0.0"
isort = "^5.12.0"
flake8-bugbear = "^23.9.16"
pep8-naming = "^0.13.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
